refactor eval environment state
